+++ egpath/src/stm32/examples/nucleo-f411re/serial/main.go
@@ -43,7 +43,7 @@ func init() {
 	periph.APB1ClockEnable(periph.USART2)
 	periph.APB1Reset(periph.USART2)
 
-	port, tx, rx := gpio.A, uint(2), uint(3)
+	port, tx, rx := gpio.A, 2, 3
 
 	port.SetMode(tx, gpio.Alt)
 	port.SetOutType(tx, gpio.PushPull)
@@ -69,7 +69,7 @@ func init() {
 	fmt.DefaultWriter = s
 }
 
-func blink(c uint, d int) {
+func blink(c int, d int) {
 	leds.SetBit(c)
 	if d > 0 {
 		delay.Millisec(d)
